# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/items/":{"get":{"summary":"Read Items","operationId":"read_items_items__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/query/{collection_id}":{"post":{"summary":"Query Database","operationId":"query_database_query__collection_id__post","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/collections/":{"get":{"tags":["collections"],"summary":"Read Collections","operationId":"read_collections_collections__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Collection"},"type":"array","title":"Response Read Collections Collections  Get"}}}}}},"post":{"tags":["collections"],"summary":"Create New Collection","operationId":"create_new_collection_collections__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/collections/{collection_id}":{"get":{"tags":["collections"],"summary":"Read Collection","operationId":"read_collection_collections__collection_id__get","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["collections"],"summary":"Update Existing Collection","operationId":"update_existing_collection_collections__collection_id__put","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["collections"],"summary":"Delete Existing Collection","operationId":"delete_existing_collection_collections__collection_id__delete","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/collections/{collection_id}/details":{"get":{"tags":["collections"],"summary":"Read Collection Details","operationId":"read_collection_details_collections__collection_id__details_get","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/tasks/{task_id}":{"delete":{"tags":["tasks"],"summary":"Delete Task","operationId":"delete_task_tasks_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/tasks":{"get":{"tags":["tasks"],"summary":"Get Tasks","operationId":"get_tasks_tasks_tasks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Response Get Tasks Tasks Tasks Get"}}}}}}},"/import/":{"get":{"tags":["import"],"summary":"Get Imports","operationId":"get_imports_import__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Import"},"type":"array","title":"Response Get Imports Import  Get"}}}}}}},"/import/{collection_id}":{"post":{"tags":["import"],"summary":"Import File","operationId":"import_file_import__collection_id__post","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_file_import__collection_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import/url/{colletion_id}":{"post":{"tags":["import"],"summary":"Import Url","operationId":"import_url_import_url__colletion_id__post","parameters":[{"name":"collection_id","in":"query","required":true,"schema":{"type":"string","title":"Collection Id"}},{"name":"url","in":"query","required":true,"schema":{"type":"string","title":"Url"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_import_url_import_url__colletion_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import/step1/{collection_id}":{"post":{"tags":["import"],"summary":"Import File Step 1","operationId":"import_file_step_1_import_step1__collection_id__post","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_file_step_1_import_step1__collection_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import/step2/{collection_id}":{"post":{"tags":["import"],"summary":"Import File Step 2","operationId":"import_file_step_2_import_step2__collection_id__post","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_import_file_step_2_import_step2__collection_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import/url/step1/{collection_id}":{"post":{"tags":["import"],"summary":"Import Url Step 1","operationId":"import_url_step_1_import_url_step1__collection_id__post","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}},{"name":"url","in":"query","required":true,"schema":{"type":"string","title":"Url"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_import_url_step_1_import_url_step1__collection_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/mcp_enabled":{"put":{"tags":["mcp"],"summary":"Set Mcp Enabled","operationId":"set_mcp_enabled_mcp_mcp_enabled_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MCPEnabledRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/logs/":{"get":{"tags":["logs"],"summary":"Read Logs","description":"Retrieve the latest n log entries.","operationId":"read_logs_logs__get","parameters":[{"name":"n","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"},"title":"Response Read Logs Logs  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/logs/stream":{"get":{"tags":["logs"],"summary":"Stream","operationId":"stream_logs_stream_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/settings/":{"get":{"tags":["settings"],"summary":"Read Settings","operationId":"read_settings_settings__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Setting"},"type":"array","title":"Response Read Settings Settings  Get"}}}}}},"put":{"tags":["settings"],"summary":"Update Settings","operationId":"update_settings_settings__put","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SettingCreate"},"type":"array","title":"Settings"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Setting"},"type":"array","title":"Response Update Settings Settings  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/{collection_id}":{"get":{"tags":["files"],"summary":"Read Files","operationId":"read_files_files__collection_id__get","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/File"},"title":"Response Read Files Files  Collection Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["files"],"summary":"Delete Files","operationId":"delete_files_files__collection_id__delete","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/content":{"post":{"tags":["files"],"summary":"Get Chunk Preview","operationId":"get_chunk_preview_files_content_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkPreviewRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkPreviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_import_file_import__collection_id__post":{"properties":{"import_params":{"type":"string","title":"Import Params"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["import_params","file"],"title":"Body_import_file_import__collection_id__post"},"Body_import_file_step_1_import_step1__collection_id__post":{"properties":{"import_params":{"type":"string","title":"Import Params"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["import_params","file"],"title":"Body_import_file_step_1_import_step1__collection_id__post"},"Body_import_file_step_2_import_step2__collection_id__post":{"properties":{"import_files_ids":{"items":{"type":"string"},"type":"array","title":"Import Files Ids"},"import_params":{"type":"string","title":"Import Params"}},"type":"object","required":["import_files_ids","import_params"],"title":"Body_import_file_step_2_import_step2__collection_id__post"},"Body_import_url_import_url__colletion_id__post":{"properties":{"import_params":{"type":"string","title":"Import Params"}},"type":"object","required":["import_params"],"title":"Body_import_url_import_url__colletion_id__post"},"Body_import_url_step_1_import_url_step1__collection_id__post":{"properties":{"import_params":{"type":"string","title":"Import Params"}},"type":"object","required":["import_params"],"title":"Body_import_url_step_1_import_url_step1__collection_id__post"},"ChunkPreviewRequest":{"properties":{"file_id":{"type":"string","title":"File Id"},"skip_number":{"type":"integer","title":"Skip Number","default":0},"take_number":{"type":"integer","title":"Take Number","default":10},"chunk_size":{"type":"integer","exclusiveMinimum":0.0,"title":"Chunk Size"},"chunk_overlap":{"type":"integer","title":"Chunk Overlap"},"no_chunks":{"type":"boolean","title":"No Chunks"}},"type":"object","required":["file_id","chunk_size","chunk_overlap","no_chunks"],"title":"ChunkPreviewRequest"},"ChunkPreviewResponse":{"properties":{"chunks":{"items":{"type":"string"},"type":"array","title":"Chunks"},"more_chunks":{"type":"boolean","title":"More Chunks"}},"type":"object","required":["chunks","more_chunks"],"title":"ChunkPreviewResponse"},"Collection":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"enabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Enabled"},"import_type":{"$ref":"#/components/schemas/ImportType","default":"NONE"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"settings":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Settings"},"id":{"type":"string","title":"Id"}},"type":"object","required":["name","id"],"title":"Collection"},"CollectionCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"enabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Enabled"},"import_type":{"$ref":"#/components/schemas/ImportType","default":"NONE"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"settings":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Settings"}},"type":"object","required":["name"],"title":"CollectionCreate"},"CollectionDetails":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"enabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Enabled"},"import_type":{"$ref":"#/components/schemas/ImportType","default":"NONE"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"settings":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Settings"},"id":{"type":"string","title":"Id"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"},"count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Count"}},"type":"object","required":["name","id"],"title":"CollectionDetails"},"File":{"properties":{"id":{"type":"string","title":"Id"},"timestamp":{"type":"string","title":"Timestamp"},"collection_id":{"type":"string","title":"Collection Id"},"path":{"type":"string","title":"Path"},"source":{"type":"string","title":"Source"}},"type":"object","required":["id","timestamp","collection_id","path","source"],"title":"File"},"FileImportSettings":{"properties":{"chunk_size":{"type":"integer","title":"Chunk Size"},"chunk_overlap":{"type":"integer","title":"Chunk Overlap"},"no_chunks":{"type":"boolean","title":"No Chunks"}},"type":"object","required":["chunk_size","chunk_overlap","no_chunks"],"title":"FileImportSettings"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Import":{"properties":{"name":{"type":"string","title":"Name"},"model":{"type":"string","title":"Model"},"settings":{"$ref":"#/components/schemas/FileImportSettings"}},"type":"object","required":["name","model","settings"],"title":"Import"},"ImportType":{"type":"string","enum":["NONE","FILE","URL"],"title":"ImportType"},"MCPEnabledRequest":{"properties":{"enabled":{"type":"boolean","title":"Enabled"}},"type":"object","required":["enabled"],"title":"MCPEnabledRequest"},"Message":{"properties":{"id":{"type":"string","title":"Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"collectionId":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Collectionid"},"topic":{"type":"string","title":"Topic"},"message":{"type":"string","title":"Message"}},"type":"object","required":["id","timestamp","topic","message"],"title":"Message"},"QueryRequest":{"properties":{"query":{"type":"string","title":"Query"},"n_results":{"type":"integer","title":"N Results","default":10}},"type":"object","required":["query"],"title":"QueryRequest"},"Setting":{"properties":{"name":{"type":"string","title":"Name"},"value":{"type":"string","title":"Value"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name","value"],"title":"Setting"},"SettingCreate":{"properties":{"name":{"type":"string","title":"Name"},"value":{"type":"string","title":"Value"}},"type":"object","required":["name","value"],"title":"SettingCreate"},"Task":{"properties":{"id":{"type":"string","title":"Id"},"collectionId":{"type":"string","title":"Collectionid"},"name":{"type":"string","title":"Name"},"startTime":{"type":"integer","title":"Starttime"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","collectionId","name","startTime","status"],"title":"Task"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}